python/converter/requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests
sqlalchemy
python-dotenv
psycopg2-binary
marker-pdf==0.3.0
more-itertools
PyPDF2